add services for Permissions API and Permission Manager plugin
